+++ x/pylons/client/cli/tx_send_items_test.go
@@ -145,12 +145,12 @@ func TestSendItems(t *testing.T) {
 
 	for i := 0; i < numTests; i++ {
 		// create execution
-		args = []string{cookbookID, recipeID, "0", "[]"} // empty list for item-ids since there is no item input
+		args = []string{cookbookID, recipeID, "0", "[]", "[]"} // empty list for item-ids since there is no item input
 		args = append(args, common...)
 		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdExecuteRecipe(), args)
 		require.NoError(t, err)
 		var resp sdk.TxResponse
-		require.NoError(t, ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &resp))
+		require.NoError(t, ctx.Codec.UnmarshalJSON(out.Bytes(), &resp))
 		require.Equal(t, uint32(0), resp.Code)
 
 		// simulate waiting for later block heights
@@ -166,7 +166,7 @@ func TestSendItems(t *testing.T) {
 		out, err = clitestutil.ExecTestCLICmd(ctx, cli.CmdShowExecution(), args)
 		require.NoError(t, err)
 		var execResp types.QueryGetExecutionResponse
-		require.NoError(t, ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &execResp))
+		require.NoError(t, ctx.Codec.UnmarshalJSON(out.Bytes(), &execResp))
 		// verify completed
 		require.Equal(t, true, execResp.Completed)
 
@@ -176,7 +176,7 @@ func TestSendItems(t *testing.T) {
 		out, err = clitestutil.ExecTestCLICmd(ctx, cli.CmdShowItem(), args)
 		require.NoError(t, err)
 		var itemResp types.QueryGetItemResponse
-		require.NoError(t, ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &itemResp))
+		require.NoError(t, ctx.Codec.UnmarshalJSON(out.Bytes(), &itemResp))
 		require.Equal(t, cookbookID, itemResp.Item.CookbookID)
 		require.Equal(t, height, itemResp.Item.LastUpdate)
 
@@ -223,7 +223,7 @@ func TestSendItems(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 				var resp sdk.TxResponse
-				require.NoError(t, ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &resp))
+				require.NoError(t, ctx.Codec.UnmarshalJSON(out.Bytes(), &resp))
 				require.Equal(t, tc.code, resp.Code)
 			}
 		})