+++ x/pylons/types/validate_common_fields_test.go
@@ -3,6 +3,8 @@ package types
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/stretchr/testify/require"
 )
 
@@ -110,20 +112,133 @@ func TestValidateUsername(t *testing.T) {
 	}
 }
 
-func TestValidateNumber(t *testing.T) {
+func TestValidatePaymentInfo(t *testing.T) {
 	for _, tc := range []struct {
 		desc string
-		id   string
+		obj  PaymentInfo
 		err  error
 	}{
-		{desc: "ValidNum", id: "0"},
-		{desc: "ValidNum1", id: "42"},
-		{desc: "InvalidNum1", id: "_1", err: ErrInvalidRequestField},
-		{desc: "InvalidNum2", id: "", err: ErrInvalidRequestField},
+		{desc: "Valid", obj: PaymentInfo{
+			PurchaseID:    "test",
+			ProcessorName: "test",
+			PayerAddr:     GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			ProductID:     "test",
+			Signature:     "test",
+		}},
+		{desc: "InvalidPurchaseID", obj: PaymentInfo{
+			PurchaseID:    "",
+			ProcessorName: "test",
+			PayerAddr:     GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			ProductID:     "test",
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidProductID", obj: PaymentInfo{
+			PurchaseID:    "test",
+			ProcessorName: "test",
+			PayerAddr:     GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			ProductID:     "",
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidSignature", obj: PaymentInfo{
+			PurchaseID:    "test",
+			ProcessorName: "test",
+			PayerAddr:     GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			ProductID:     "test",
+			Signature:     "",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidAmount", obj: PaymentInfo{
+			PurchaseID:    "test",
+			ProcessorName: "test",
+			PayerAddr:     GenTestBech32FromString("test"),
+			Amount:        sdk.ZeroInt(),
+			ProductID:     "test",
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidAddress", obj: PaymentInfo{
+			PurchaseID:    "test",
+			ProcessorName: "test",
+			PayerAddr:     "test",
+			Amount:        sdk.OneInt(),
+			ProductID:     "test",
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidProcessorName", obj: PaymentInfo{
+			PurchaseID:    "test",
+			ProcessorName: "",
+			PayerAddr:     GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			ProductID:     "test",
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
 	} {
 		tc := tc
 		t.Run(tc.desc, func(t *testing.T) {
-			err := ValidateNumber(tc.id)
+			err := ValidatePaymentInfo(tc.obj)
+			if tc.err != nil {
+				require.ErrorIs(t, err, tc.err)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestValidateRedeemInfo(t *testing.T) {
+	for _, tc := range []struct {
+		desc string
+		obj  RedeemInfo
+		err  error
+	}{
+		{desc: "Valid", obj: RedeemInfo{
+			ID:            "test",
+			ProcessorName: "test",
+			Address:       GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			Signature:     "test",
+		}},
+		{desc: "InvalidID", obj: RedeemInfo{
+			ID:            "",
+			ProcessorName: "test",
+			Address:       GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidSignature", obj: RedeemInfo{
+			ID:            "test",
+			ProcessorName: "test",
+			Address:       GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			Signature:     "",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidAmount", obj: RedeemInfo{
+			ID:            "test",
+			ProcessorName: "test",
+			Address:       GenTestBech32FromString("test"),
+			Amount:        sdk.ZeroInt(),
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidAddress", obj: RedeemInfo{
+			ID:            "test",
+			ProcessorName: "test",
+			Address:       "test",
+			Amount:        sdk.OneInt(),
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+		{desc: "InvalidProcessorName", obj: RedeemInfo{
+			ID:            "test",
+			ProcessorName: "",
+			Address:       GenTestBech32FromString("test"),
+			Amount:        sdk.OneInt(),
+			Signature:     "test",
+		}, err: ErrInvalidRequestField},
+	} {
+		tc := tc
+		t.Run(tc.desc, func(t *testing.T) {
+			err := ValidateRedeemInfo(tc.obj)
 			if tc.err != nil {
 				require.ErrorIs(t, err, tc.err)
 			} else {